+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/Entities.hs
@@ -20,6 +20,8 @@ module Cardano.Ledger.Dijkstra.Rules.Entities (
   EntitiesEvent (..),
   validateWrongNetworkInDirectDeposit,
   validateMissingAccountsInDirectDeposits,
+  validateMissingAccountsInWithdrawals,
+  validateMissingOriginalAccountsInWithdrawals,
 ) where
 
 import Cardano.Ledger.Address (DirectDeposits (..))
@@ -27,51 +29,80 @@ import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Coin (Coin)
+import Cardano.Ledger.Compactible
 import Cardano.Ledger.Conway.Core
+import Cardano.Ledger.Conway.Governance (
+  Committee,
+  GovActionPurpose (..),
+  GovActionState,
+  GovPurposeId,
+ )
 import qualified Cardano.Ledger.Conway.Rules as Conway
 import Cardano.Ledger.Conway.State
 import Cardano.Ledger.Dijkstra.Era (DijkstraEra, ENTITIES)
 import Cardano.Ledger.Dijkstra.Rules.Certs ()
 import Cardano.Ledger.Dijkstra.Rules.GovCert (DijkstraGovCertPredFailure)
-import Cardano.Ledger.Dijkstra.TxBody (DijkstraEraTxBody, directDepositsTxBodyL)
+import Cardano.Ledger.Dijkstra.TxBody
+import Cardano.Ledger.Dijkstra.UTxO (DijkstraEraUTxO (..))
 import Cardano.Ledger.Rules.ValidationMode (Test, runTest)
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
 import Control.DeepSeq (NFData)
 import Control.Monad.Trans.Reader (asks)
 import Control.State.Transition.Extended
+import Data.Foldable (foldMap')
 import Data.Map.NonEmpty (NonEmptyMap)
-import qualified Data.Map.NonEmpty as NEM
 import qualified Data.Map.Strict as Map
 import Data.Sequence (Seq)
+import qualified Data.Sequence.Strict as StrictSeq
 import Data.Set.NonEmpty (NonEmptySet)
 import GHC.Generics (Generic)
 import Lens.Micro
 
 data EntitiesEnv era = EntitiesEnv
-  { eePlutusLegacyMode :: Bool
-  , eeCertsEnv :: Conway.CertsEnv era
+  { eeCurrentEpoch :: EpochNo
+  , eePParams :: PParams era
+  , eeCurrentCommittee :: StrictMaybe (Committee era)
+  , eeCommitteeProposals :: Map.Map (GovPurposeId 'CommitteePurpose) (GovActionState era)
+  , eeOriginalAccounts :: Accounts era
   }
   deriving (Generic)
 
-deriving instance (EraPParams era, Eq (Tx TopTx era)) => Eq (EntitiesEnv era)
+deriving instance
+  (EraPParams era, Eq (Committee era), Eq (GovActionState era), Eq (Accounts era)) =>
+  Eq (EntitiesEnv era)
 
-deriving instance (EraPParams era, Show (Tx TopTx era)) => Show (EntitiesEnv era)
+deriving instance
+  (EraPParams era, Show (Committee era), Show (GovActionState era), Show (Accounts era)) =>
+  Show (EntitiesEnv era)
 
-instance (EraPParams era, NFData (Tx TopTx era)) => NFData (EntitiesEnv era)
+instance
+  (EraPParams era, NFData (Committee era), NFData (GovActionState era), NFData (Accounts era)) =>
+  NFData (EntitiesEnv era)
 
-instance EraTx era => EncCBOR (EntitiesEnv era) where
-  encCBOR x@(EntitiesEnv _ _) =
+instance
+  ( EraPParams era
+  , EncCBOR (Committee era)
+  , EncCBOR (GovActionState era)
+  , EncCBOR (Accounts era)
+  ) =>
+  EncCBOR (EntitiesEnv era)
+  where
+  encCBOR x@(EntitiesEnv _ _ _ _ _) =
     let EntitiesEnv {..} = x
      in encode $
           Rec EntitiesEnv
-            !> To eePlutusLegacyMode
-            !> To eeCertsEnv
+            !> To eeCurrentEpoch
+            !> To eePParams
+            !> To eeCurrentCommittee
+            !> To eeCommitteeProposals
+            !> To eeOriginalAccounts
 
 data EntitiesPredFailure era
   = CertsFailure (PredicateFailure (EraRule "CERTS" era))
   | MissingAccountsInWithdrawals Withdrawals
-  | IncompleteWithdrawals (NonEmptyMap AccountAddress (Mismatch RelEQ Coin))
-  | ExceededBalancesInWithdrawals (NonEmptyMap AccountAddress (Mismatch RelLTEQ Coin))
+  | MissingOriginalAccountsInWithdrawals Withdrawals
+  | InexactWithdrawalsInLegacy (NonEmptyMap AccountAddress (Mismatch RelEQ Coin))
+  | WithdrawalsExceedAccountBalance (NonEmptyMap AccountAddress (Mismatch RelLTEQ Coin))
   | MissingAccountsInDirectDeposits DirectDeposits
   | WrongNetworkInWithdrawals
       -- | Expected network id
@@ -108,11 +139,12 @@ instance
     encode . \case
       CertsFailure x -> Sum (CertsFailure @era) 0 !> To x
       MissingAccountsInWithdrawals x -> Sum (MissingAccountsInWithdrawals @era) 1 !> To x
-      IncompleteWithdrawals x -> Sum (IncompleteWithdrawals @era) 2 !> To x
-      ExceededBalancesInWithdrawals x -> Sum (ExceededBalancesInWithdrawals @era) 3 !> To x
+      InexactWithdrawalsInLegacy x -> Sum (InexactWithdrawalsInLegacy @era) 2 !> To x
+      WithdrawalsExceedAccountBalance x -> Sum (WithdrawalsExceedAccountBalance @era) 3 !> To x
       MissingAccountsInDirectDeposits x -> Sum (MissingAccountsInDirectDeposits @era) 4 !> To x
       WrongNetworkInWithdrawals x y -> Sum (WrongNetworkInWithdrawals @era) 5 !> To x !> To y
       WrongNetworkInDirectDeposits x y -> Sum (WrongNetworkInDirectDeposits @era) 6 !> To x !> To y
+      MissingOriginalAccountsInWithdrawals x -> Sum (MissingOriginalAccountsInWithdrawals @era) 7 !> To x
 
 instance
   ( Era era
@@ -123,11 +155,12 @@ instance
   decCBOR = decode . Summands "EntitiesPredFailure" $ \case
     0 -> SumD CertsFailure <! From
     1 -> SumD MissingAccountsInWithdrawals <! From
-    2 -> SumD IncompleteWithdrawals <! From
-    3 -> SumD ExceededBalancesInWithdrawals <! From
+    2 -> SumD InexactWithdrawalsInLegacy <! From
+    3 -> SumD WithdrawalsExceedAccountBalance <! From
     4 -> SumD MissingAccountsInDirectDeposits <! From
     5 -> SumD WrongNetworkInWithdrawals <! From <! From
     6 -> SumD WrongNetworkInDirectDeposits <! From <! From
+    7 -> SumD MissingOriginalAccountsInWithdrawals <! From
     n -> Invalid n
 
 newtype EntitiesEvent era = CertsEvent (Event (EraRule "CERTS" era))
@@ -170,6 +203,7 @@ instance InjectRuleFailure "ENTITIES" Shelley.ShelleyUtxoPredFailure DijkstraEra
 instance
   ( EraTx era
   , DijkstraEraTxBody era
+  , DijkstraEraUTxO era
   , ConwayEraPParams era
   , ConwayEraCertState era
   , Embed (EraRule "CERTS" era) (ENTITIES era)
@@ -184,7 +218,7 @@ instance
   STS (ENTITIES era)
   where
   type State (ENTITIES era) = CertState era
-  type Signal (ENTITIES era) = Seq (TxCert era)
+  type Signal (ENTITIES era) = StAnnTx TopTx era
   type Environment (ENTITIES era) = EntitiesEnv era
   type BaseM (ENTITIES era) = ShelleyBase
   type PredicateFailure (ENTITIES era) = EntitiesPredFailure era
@@ -195,8 +229,8 @@ instance
 
 dijkstraEntitiesTransition ::
   forall era.
-  ( EraTx era
-  , DijkstraEraTxBody era
+  ( DijkstraEraTxBody era
+  , DijkstraEraUTxO era
   , ConwayEraCertState era
   , Embed (EraRule "CERTS" era) (ENTITIES era)
   , State (EraRule "CERTS" era) ~ CertState era
@@ -209,17 +243,32 @@ dijkstraEntitiesTransition ::
   ) =>
   TransitionRule (ENTITIES era)
 dijkstraEntitiesTransition = do
-  TRC (EntitiesEnv legacyMode certsEnv, certState, certificates) <- judgmentContext
-  let Conway.CertsEnv tx pp curEpoch _committee _committeeProposals = certsEnv
+  TRC (EntitiesEnv curEpoch pp committee committeeProposals originalAccounts, certState, stAnnTx) <-
+    judgmentContext
+  let tx = stAnnTx ^. txStAnnTxG
+      legacyMode = stAnnTx ^. plutusLegacyModeStAnnTxG
       withdrawals = tx ^. bodyTxL . withdrawalsTxBodyL
+      subTxWithdrawals =
+        foldMap'
+          (\subTx -> subTx ^. bodyTxL . withdrawalsTxBodyL)
+          (tx ^. bodyTxL . subTransactionsTxBodyL)
+      batchWithdrawals = withdrawals <> subTxWithdrawals
       accounts = certState ^. certDStateL . accountsL
+      certsEnv = Conway.CertsEnv tx pp curEpoch committee committeeProposals
 
   network <- liftSTS $ asks networkId
 
   runTest $ Shelley.validateWrongNetworkWithdrawal network (tx ^. bodyTxL)
   runTest $ validateWrongNetworkInDirectDeposit network (tx ^. bodyTxL)
+  runTest $ validateMissingAccountsInWithdrawals withdrawals accounts
 
-  validateWithdrawals legacyMode network withdrawals accounts
+  if legacyMode
+    then do
+      runTest $ validateIncompleteWithdrawals withdrawals network accounts
+      runTest $ validateExceededBalancesInWithdrawals subTxWithdrawals network originalAccounts
+    else do
+      runTest $ validateMissingOriginalAccountsInWithdrawals withdrawals originalAccounts
+      runTest $ validateExceededBalancesInWithdrawals batchWithdrawals network originalAccounts
 
   let certStateBeforeCerts =
         certState
@@ -227,7 +276,8 @@ dijkstraEntitiesTransition = do
           & Conway.updateVotingDRepExpiries tx curEpoch (pp ^. ppDRepActivityL)
           & certDStateL . accountsL %~ applyWithdrawals withdrawals
   certStateAfterCerts <-
-    trans @(EraRule "CERTS" era) $ TRC (certsEnv, certStateBeforeCerts, certificates)
+    trans @(EraRule "CERTS" era) $
+      TRC (certsEnv, certStateBeforeCerts, StrictSeq.fromStrict $ tx ^. bodyTxL . certsTxBodyL)
 
   let directDeposits = tx ^. bodyTxL . directDepositsTxBodyL
       accountsAfterCerts = certStateAfterCerts ^. certDStateL . accountsL
@@ -259,32 +309,61 @@ validateMissingAccountsInDirectDeposits dds accounts =
     (directDepositsMissingAccounts dds accounts)
     MissingAccountsInDirectDeposits
 
-validateWithdrawals ::
+validateMissingAccountsInWithdrawals ::
   EraAccounts era =>
-  Bool ->
-  Network ->
   Withdrawals ->
   Accounts era ->
-  Rule (ENTITIES era) ctx ()
-validateWithdrawals legacyMode network withdrawals accounts = do
-  missingWithdrawals <-
-    if legacyMode
-      then do
-        let (missingWithdrawals, incompleteWithdrawals) =
-              case withdrawalsThatDoNotDrainAccounts withdrawals network accounts of
-                Nothing -> (Map.empty, Map.empty)
-                Just (missing, incomplete) -> (unWithdrawals missing, incomplete)
-        failOnNonEmptyMap incompleteWithdrawals IncompleteWithdrawals
-        pure missingWithdrawals
-      else do
-        let (missingWithdrawals, exceededWithdrawals) =
-              case withdrawalsThatExceedAccountBalance withdrawals network accounts of
-                Nothing -> (Map.empty, Map.empty)
-                Just (missing, exceeded) -> (unWithdrawals missing, exceeded)
-        failOnNonEmptyMap exceededWithdrawals ExceededBalancesInWithdrawals
-        pure missingWithdrawals
-  failOnNonEmptyMap missingWithdrawals $
-    MissingAccountsInWithdrawals . Withdrawals . NEM.toMap
+  Test (EntitiesPredFailure era)
+validateMissingAccountsInWithdrawals wdrls accounts =
+  failureOnJust
+    (withdrawalsMissingAccounts wdrls accounts)
+    MissingAccountsInWithdrawals
+
+validateMissingOriginalAccountsInWithdrawals ::
+  EraAccounts era =>
+  Withdrawals ->
+  Accounts era ->
+  Test (EntitiesPredFailure era)
+validateMissingOriginalAccountsInWithdrawals withdrawals originalAccounts =
+  failureOnJust
+    (withdrawalsMissingAccounts withdrawals originalAccounts)
+    MissingOriginalAccountsInWithdrawals
+
+validateExceededBalancesInWithdrawals ::
+  EraAccounts era =>
+  Withdrawals ->
+  Network ->
+  Accounts era ->
+  Test (EntitiesPredFailure era)
+validateExceededBalancesInWithdrawals withdrawals networkId accounts =
+  failureOnJust
+    ( withdrawalsWithUnacceptableAmount
+        ( \withdrawalAmount account ->
+            withdrawalAmount <= fromCompact (account ^. balanceAccountStateL)
+        )
+        withdrawals
+        networkId
+        accounts
+    )
+    WithdrawalsExceedAccountBalance
+
+validateIncompleteWithdrawals ::
+  EraAccounts era =>
+  Withdrawals ->
+  Network ->
+  Accounts era ->
+  Test (EntitiesPredFailure era)
+validateIncompleteWithdrawals withdrawals networkId accounts =
+  failureOnJust
+    ( withdrawalsWithUnacceptableAmount
+        ( \withdrawalAmount account ->
+            withdrawalAmount == fromCompact (account ^. balanceAccountStateL)
+        )
+        withdrawals
+        networkId
+        accounts
+    )
+    InexactWithdrawalsInLegacy
 
 conwayToDijkstraEntitiesPredFailure ::
   forall era. Conway.ConwayLedgerPredFailure era -> EntitiesPredFailure era